Added the patch from Nicholas Redgrave to the CVS:
Greetings all.

I used MPlayer to rip an MPEG2 file from a DVD-RAM video disc into separate
WAV and video-only AVI files.
The AVI file was 1.5 GB and the WAV file 398 MB.
I wanted to use lavaddwav to multiplex them back together into a file that I
could edit with lavplay but it stopped when it reached the 2 GB avi file
limit.
So I have attempted to modify the lavaddwav code so that you can use an output
file format style of "outfile%02d.avi" to create sequential files where it
continues writing to the next file when the first reaches about 1.7 GB.

Hopefully there is a patch file attached.


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code/trunk/mjpeg_play@2622 3be20b7b-d6de-45d7-9b51-1db703c412de