The deprecated XYSCSS tag for chroma (offical tag for 1.7.x is 'C') was causing a problem with programs that ran at elevated "feature" level (such as y4mscaler).  The 'C' tag would be seen and believed but the old 'X' tag not removed.  This lead to the problem of headers like this:
YUV4MPEG2 W720 H480 F30000:1001 Ib A10:11 C420mpeg2 XYSCSS=411

Note the 'C' of 420mpeg2 and 'XYSCSS' of 411!

At any rate the old X tag had to be supported until a release is made.  At
that time ffmpeg/mplayer/smilutils can be modified to only emit 'C' rather
than both the 'C' and 'X' tag.

This change believes the 'X' tag (since the C and X had to be the same on
input) and then DELETES so that it does not escape back into the world.


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code/trunk/mjpeg_play@2624 3be20b7b-d6de-45d7-9b51-1db703c412de