Added a patch from the tracker submitted by: Dean Gaudet [ 1117718 ] jpeg2yuv: eliminate division in rescale_color_vals
rescale_color_vals includes a division and a multiplication, but it can be refactored so the division can be done at compile time. (compilers are generally not permitted to make this change, but most have compile options which loosen the fp standards compliance which would catch this case.)

in a pipeline with mpeg2enc this yielded a 9% speed improvement.


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code/trunk/mjpeg_play@2679 3be20b7b-d6de-45d7-9b51-1db703c412de