Finally fixed the earthquaking problem in dark/cloudy areas!
The problem was that the match-size-throttle was picking a bad match
before it ever had a chance to find a good match.

Now, the match-size-throttle isn't applied unless the
match-count-throttle is also exceeded, or the end of a line is reached.

This has led to a SERIOUS quality increase, without making y4mdenoise
any slower than it was before!


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code/trunk/mjpeg_play@3279 3be20b7b-d6de-45d7-9b51-1db703c412de